used static casts instead of implicit casts for signal handler types

+++ src/encoder_interface.cpp
@@ -63,9 +63,9 @@ void EncoderInterface::handle_all(const can_frame& message) {
 void EncoderInterface::handle_angle(const uint8_t* message_data, const uint32_t can_id) {
     if (message_data[0] == 0x55 && message_data[1] == 0x55) {
         uint16_t angle_register_value = static_cast<uint16_t>(message_data[3] << 8) | message_data[2];
-        uint16_t angle = angle_register_value * 360.0 / 32768;
+        uint16_t angle = static_cast<uint16_t>(angle_register_value * 360.0 / 32768);
         uint16_t angular_velocity_register_value = static_cast<uint16_t>(message_data[5] << 8) | message_data[4];
-        uint16_t angular_velocity = angular_velocity_register_value * 360.0 / 32768 / ANGULAR_VELOCITY_SAMPLE_TIME;
+        uint16_t angular_velocity = static_cast<uint16_t>(angular_velocity_register_value * 360.0 / 32768 / ANGULAR_VELOCITY_SAMPLE_TIME);
         uint16_t number_of_rotations = static_cast<uint16_t>(message_data[7] << 8) | message_data[6];
         angle_signal(can_id, angle, angular_velocity, number_of_rotations);
     }
@@ -74,7 +74,7 @@ void EncoderInterface::handle_angle(const uint8_t* message_data, const uint32_t 
 void EncoderInterface::handle_temp(const uint8_t* message_data, const uint32_t can_id) {
     if (message_data[0] == 0x55 && message_data[1] == 0x56) {
         uint16_t temperature_register_value = static_cast<uint16_t>(message_data[3] << 8) | message_data[2];
-        uint16_t temperature = temperature_register_value / 100;
+        uint16_t temperature = static_cast<uint16_t>(temperature_register_value / 100);
         temp_signal(can_id, temperature);
     }
 };
